+++ archive/tar_unix.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package archive
 
 import (
