+++ archive/time_unix.go
@@ -1,3 +1,5 @@
+// +build linux freebsd
+
 package archive
 
 import (
