+++ cmd/containerd/builtins_unix.go
@@ -1,0 +1,7 @@
+// +build darwin freebsd
+
+package main
+
+import (
+	_ "github.com/containerd/containerd/snapshot/naive"
+)
