+++ cmd/ctr/shim.go
@@ -1,4 +1,4 @@
-// +build linux
+// +build !windows
 
 package main
