+++ content/store_linux.go
@@ -1,5 +1,3 @@
-// +build darwin freebsd
-
 package content
 
 import (
@@ -10,7 +8,7 @@ import (
 
 func getStartTime(fi os.FileInfo) time.Time {
 	if st, ok := fi.Sys().(*syscall.Stat_t); ok {
-		return time.Unix(int64(st.Ctimespec.Sec), int64(st.Ctimespec.Nsec))
+		return time.Unix(int64(st.Ctim.Sec), int64(st.Ctim.Nsec))
 	}
 
 	return fi.ModTime()
