+++ vendor.conf
@@ -20,7 +20,7 @@ github.com/containerd/btrfs e9c546f46bccffefe71a6bc137e4c21b5503cc18
 github.com/stretchr/testify v1.1.4
 github.com/davecgh/go-spew v1.1.0
 github.com/pmezard/go-difflib v1.0.0
-github.com/tonistiigi/fifo fe870ccf293940774c2b44e23f6c71fff8f7547d
+github.com/tonistiigi/fifo f071cd4a2739654fec4e768a14efd9332b3e8af1
 github.com/urfave/cli 8ba6f23b6e36d03666a14bd9421f5e3efcb59aca
 golang.org/x/net 8b4af36cd21a1f85a7484b49feb7c79363106d8e
 google.golang.org/grpc v1.0.5
